+++ src/System/CommandExecutor.php
@@ -11,6 +11,8 @@
 
 namespace Plista\Chimney\System;
 
+use Plista\Chimney\Command\Make\ExitException;
+
 /**
  *
  */
@@ -21,10 +23,25 @@ class CommandExecutor implements ExecutorInterface
      */
     public function execute($program, $parameters='')
     {
-        exec(
-            $parameters ? "{$program} $parameters" : $program,
-            $output
+        exec($parameters ? "{$program} $parameters" : $program,
+            $output,
+            $returnVar
         );
+
+        /** Adding an ExitException when an error is encountered (i.e., $returnVar > 0)
+          * The reason for the change is that if the post-run script encounters an error,
+          * chimney doesn't forward it to the caller which causes problems with
+          * gitlab-ci marking jobs incorrectly as successful.
+        */
+        if ($returnVar > 0) {
+            $outputStr = implode("\n", $output);
+
+            throw new ExitException(
+                "The command below returned non-zero value: $returnVar \n$outputStr",
+                ExitException::STATUS_ILLEGAL_COMMAND
+            );
+        }
+
         return $output;
     }
 }